# demo/safe/swagger.json
{
  "openapi": "3.0.1",
  "servers": [
    {
      "url": "https://github.com/samchon/nestia",
      "description": "insert your server url"
    }
  ],
  "info": {
    "version": "0.1.0",
    "title": "Generated by nestia - https://github.com/samchon/nestia"
  },
  "paths": {
    "/consumers/{section}/sales/{saleId}/articles/{articleId}/comments": {
      "post": {
        "tags": [],
        "parameters": [
          {
            "name": "section",
            "in": "path",
            "description": "Code of the target section",
            "schema": {
              "type": "string",
              "nullable": false
            },
            "required": true
          },
          {
            "name": "saleId",
            "in": "path",
            "description": "ID of the target sale",
            "schema": {
              "type": "string",
              "nullable": false
            },
            "required": true
          },
          {
            "name": "articleId",
            "in": "path",
            "description": "ID of the target article",
            "schema": {
              "type": "string",
              "nullable": false
            },
            "required": true
          }
        ],
        "requestBody": {
          "description": "Content to write",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ISaleArticleComment.IStore"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Newly archived comment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ISaleArticleComment"
                }
              }
            }
          },
          "400": {
            "description": "bad request error when type of the input data is not valid"
          },
          "401": {
            "description": "unauthorized error when you've not logged in yet"
          },
          "403": {
            "description": "forbidden error when you're a seller and the sale is not yours"
          },
          "404": {
            "description": "not found error when unable to find the matched record"
          }
        },
        "description": "Store a new comment.\n\nWrite a comment on a sale article. If you configure the comment to be\n`anonymous`, only administrator, you and seller of the sale can read\nthe content."
      }
    }
  },
  "components": {
    "schemas": {
      "ISaleArticleComment.IStore": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string",
            "nullable": false,
            "description": "Body of the content."
          },
          "extension": {
            "type": "string",
            "enum": [
              "html",
              "md",
              "txt"
            ],
            "nullable": false,
            "description": "Extension, content type (format)."
          },
          "annonymous": {
            "type": "boolean",
            "nullable": false,
            "description": "Whether to hide the writer name or not."
          }
        },
        "nullable": false,
        "required": [
          "body",
          "extension",
          "annonymous"
        ],
        "description": "Store info."
      },
      "ISaleArticleComment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "nullable": false,
            "description": "Primary Key."
          },
          "parent_id": {
            "type": "number",
            "nullable": true,
            "description": "Parent comment ID.\n\nOnly When this comment has been written as a reply."
          },
          "writer_type": {
            "type": "string",
            "enum": [
              "seller",
              "consumer"
            ],
            "nullable": false,
            "description": "Type of the writer."
          },
          "writer_name": {
            "type": "string",
            "nullable": true,
            "description": "Name of the writer.\n\nWhen this is a type of anonymous comment, writer name would be hidden."
          },
          "contents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ISaleArticleComment.IContent",
              "description": "Contents of the comments.\n\nWhen the comment writer tries to modify content, it would not modify the comment\ncontent but would be accumulated. Therefore, all of the people can read how\nthe content has been changed."
            },
            "nullable": false,
            "description": "Contents of the comments.\n\nWhen the comment writer tries to modify content, it would not modify the comment\ncontent but would be accumulated. Therefore, all of the people can read how\nthe content has been changed."
          },
          "created_at": {
            "type": "string",
            "nullable": false,
            "description": "Creation time."
          }
        },
        "nullable": false,
        "required": [
          "id",
          "parent_id",
          "writer_type",
          "writer_name",
          "contents",
          "created_at"
        ],
        "description": "Comment wrote on a sale related article.\n\nWhen an article of a sale has been enrolled, all of the participants like consumers and\nsellers can write a comment on that article. However, when the writer is a consumer, the\nconsumer can hide its name through the annoymous option.\n\nAlso, writing a reply comment for a specific comment is possible and in that case, the\n{@link ISaleArticleComment.parent_id} property would be activated."
      },
      "ISaleArticleComment.IContent": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": false,
            "description": "Primary Key."
          },
          "body": {
            "type": "string",
            "nullable": false,
            "description": "Body of the content."
          },
          "created_at": {
            "type": "string",
            "nullable": false,
            "description": "Creation time."
          }
        },
        "nullable": false,
        "required": [
          "id",
          "body",
          "created_at"
        ],
        "description": "Content info."
      }
    }
  }
}